fully functional

+++ src/frontend/WorkSpace.java
@@ -8,7 +8,7 @@ import backend.Variable;
 
 public class WorkSpace implements java.io.Serializable {
 	public String language;
-	public int background;
+	public double background;
 	public List<String> views;
 	public Map<Double, String> colorPalette;
 	public Map<Double, String> imagePalette;
